+++ Source/Cells/DatePickerCell.swift
@@ -10,7 +10,6 @@ struct DatePickerCellConstants {
 
 public class DatePickerCellModel {
 	var title: String = ""
-	var toolbarMode: ToolbarMode = .Simple
 	var datePickerMode: UIDatePickerMode = .DateAndTime
 	var locale: NSLocale? = nil // default is [NSLocale currentLocale]. setting nil returns to default
 	var minimumDate: NSDate? = nil // specify min/max date range. default is nil. When min > max, the values are ignored. Ignored in countdown timer mode